requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dotenv==1.0.0
